Refactor DDD, cache invalidation via publisher, UX for frontend

+++ src/Application/DTOs/BoardLists/BoardListResponse.cs
@@ -5,6 +5,6 @@ public sealed record BoardListResponse
     public Guid Id { get; set; }
     public Guid BoardId { get; set; }
     public string Title { get; set; } = string.Empty;
-    public int Position { get; set; }
+    public double Position { get; set; }
     public int ColorArgb { get; set; }
 }
